Fix forum admin stats
